requirements.txt
gitpython
tree-sitter==0.21.3
tree-sitter-languages
typer[all]
rich
pytest
